Refactor process handling and improved error handling in file hashing
Changed return type of a function to always return a Process object to avoid null results in the src/functions.php. An echo command Process object is now returned instead of false when needing to skip due to fingerprint match.
Refactor the fingerprint-and-process-handling method in the examples/fingerprint.php for better readability and clarity.
Unnecessary imports removed from tests/Fingerprint/FingerprintSimpleTaskTest.php.
Improvements were made in the src/HasherHelper.php for file listing and hashing. Exception handling was added for invalid glob patterns and the modified time fetch for the MTimes strategy.
Also, the order of methods in the src/Fingerprint/FingerprintHelper.php was changed for better structure.
In tests, the order of methods in tests/Fingerprint/FingerprintedTest.php was also switched for structured code.
Tests and Examples files were refractor for clear and clean code along with removal of unnecessary whitespaces.