Remove fingerprint from AsTask attribute and improve fingerprint usage
The commit removes the use of callable fingerprints in the AsTask attribute and improves their usage in tasks by adding dedicated functions to manage fingerprints. The reason for this change is to simplify the process of managing fingerprints and make its usage clearer in tasks. The new dedicated functions, namely fingerprint_exists, fingerprint and fingerprint_save provide a more streamlined way of managing fingerprints.

Additionally, the commit updates the tests and documentation to reflect these changes. Code cleanup was also performed within functions.php, improving readability.

The commit also removes the force option, which allowed tasks to be forcibly executed. This removal was made inline the changes to fingerprint functionality.